MOVE to functional/type_classes
